add ext files for each topics and small refactoring
