+++ contest/src/main/java/com/github/contest/array/ArrayExt.kt
@@ -1,0 +1,16 @@
+package com.github.contest.array
+
+fun IntArray.printArray() {
+    val s = when (this.size) {
+        0 -> "[]"
+        1 -> "[${this[0]}]"
+        2 -> "[${this[0]}, ${this[1]}]"
+        else -> {
+            var temp = "[${this[0]}, "
+            for (i in 1 until this.size - 1) temp += "${this[i]}, "
+            temp += "${this[this.size - 1]}]"
+            temp
+        }
+    }
+    println(s)
+}
