+++ contest/src/main/java/com/github/contest/hashTable/MapExt.kt
@@ -1,5 +1,4 @@
-package com.github.contest
-
+package com.github.contest.hashTable
 
 fun <K, V> MutableMap<K, V>.removeIfEmptyBucket(key: K) {
     if (this[key] == 0) this.remove(key)
@@ -8,10 +7,4 @@ fun <K, V> MutableMap<K, V>.removeIfEmptyBucket(key: K) {
 fun <K> MutableMap<K, Int>.reduceCount(key: K) {
     this[key] = this.getOrDefault(key, 0) - 1
     removeIfEmptyBucket(key)
-}
-
-
-fun abs(number: Int): Int = when {
-    number < 0 -> number * -1
-    else -> number
 }
