+++ contest/src/main/java/com/github/contest/math/MathExt.kt
@@ -1,0 +1,6 @@
+package com.github.contest.math
+
+fun abs(number: Int): Int = when {
+    number < 0 -> number * -1
+    else -> number
+}
